projects/open-audit/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openaudit"
version = "0.1.0"
description = "Security & bias auditor for AI agents"
dependencies = [
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "jsonschema>=4.0",
    "jinja2>=3.0",
    "click>=8.0",
    "pytest>=7.0",
]

[project.scripts]
openaudit = "openaudit.cli:main"

[tool.black]
line-length = 100

[tool.ruff]
select = ["E", "F", "W"]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[tool.hatch.build]
include = [
    "src/openaudit/py.typed",
    "src/openaudit/data/*.yaml",
]
exclude = []
